token auth for methods

+++ controller/comments.js
@@ -2,6 +2,7 @@
 
 var response = require('../res');
 var connection = require('../conn');
+var header = require('../header');
 
 exports.readComments = function(req, res) {
     var project_id = req.params.project_id;
@@ -21,37 +22,43 @@ exports.readComments = function(req, res) {
 };
 
 exports.createComments = function(req, res) {
+    if(!req.headers.authorization){
+        response.credErr('Unauthorized', res);
+    }else{
+        var comment = req.body.comment;
+        var user_id = req.body.user_id;
+        var project_id = req.body.project_id;
 
-    var comment = req.body.comment;
-    var user_id = req.body.user_id;
-    var project_id = req.body.project_id;
-
-    try{
-        connection.query('INSERT INTO Comments (comment, user_id, project_id) values (?,?,?)',
-        [ comment, user_id, project_id ],
-        function (error, rows, fields){
-            if(error){
-                response.internalError(error.code, res);
-            } else{
-                response.ok("Berhasil menambahkan user!", res);
-            }
-        });
-    }catch(err){
-        response.clientError("Bad Request", res);
+        try{
+            connection.query('INSERT INTO Comments (comment, user_id, project_id) values (?,?,?)',
+            [ comment, user_id, project_id ],
+            function (error, rows, fields){
+                if(error){
+                    response.internalError(error.code, res);
+                } else{
+                    response.ok("Berhasil menambahkan user!", res);
+                }
+            });
+        }catch(err){
+            response.clientError("Bad Request", res);
+        }
     }
 };
 
 exports.deleteComments = function(req, res) {
+    if(!req.headers.authorization){
+        response.credErr('Unauthorized', res);
+    }else{
+        var id = req.params.id;
 
-    var id = req.params.id;
-
-    connection.query('DELETE FROM Comments WHERE id = ?',
-    [ id ],
-    function (error, rows, fields){
-        if(error){
-            console.log(error)
-        } else{
-            response.ok("Berhasil menghapus user!", res)
-        }
-    });
+        connection.query('DELETE FROM Comments WHERE id = ?',
+        [ id ],
+        function (error, rows, fields){
+            if(error){
+                console.log(error)
+            } else{
+                response.ok("Berhasil menghapus user!", res)
+            }
+        });
+    }
 };
